+++ src/main/kotlin/com/github/niclasvaneyk/laravelmake/plugin/jetbrains/LaravelIcons.kt
@@ -4,6 +4,7 @@ import com.intellij.openapi.util.IconLoader
 
 object LaravelIcons {
     val LaravelLogo = IconLoader.getIcon("/icons/laravel-logo.svg", LaravelIcons::class.java)
+    val LaravelLogoForToolWindowForNewUI = IconLoader.getIcon("/icons/laravel-logo-for-new-tool-window.svg", LaravelIcons::class.java)
     val RouteForGutter = IconLoader.getIcon("/icons/route-gutter.svg", LaravelIcons::class.java)
     val Sail = IconLoader.getIcon("/icons/sail.svg", LaravelIcons::class.java)
 }